infrastructure/requirements.txt
aws-cdk-lib==2.253.0
constructs==10.6.0
